chore(web-app): move asset lp images to pages dir
