--- services/web-app/pages/assets/asset-landing-page/index.js
@@ -1,7 +1,0 @@
-/*
- * Copyright IBM Corp. 2021, 2022
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-export { default } from './asset-landing-page'
