+++ services/web-app/pages/assets/index/index.module.scss
